fix: rename datePicker case sensitive
