+++ src/components/DatePicker/index.android.js
@@ -5,7 +5,7 @@ import moment from 'moment';
 import _ from 'underscore';
 import TextInput from '../TextInput';
 import CONST from '../../CONST';
-import {propTypes, defaultProps} from './datepickerPropTypes';
+import {propTypes, defaultProps} from './datePickerPropTypes';
 import styles from '../../styles/styles';
 
 class DatePicker extends React.Component {
